Refactor learning modules and replace rule engine with suggestion engines
Reorganized learning modules into separate submodules for better structure and clarity. Replaced `RuleEngine` with `SuggestionEngine` and introduced LLM and rule-based suggestion implementations. Simplified transaction suggestion logic in the API layer with flexible input and output handling.
