+++ fintrack-api/build.gradle.kts
@@ -71,7 +71,8 @@ dependencies {
     implementation(project(":core"))
     implementation(project(":domain"))
     implementation(project(":jpa-repository"))
-    implementation(project(":learning:rule-engine"))
+    implementation(project(":learning:learning-module"))
+    implementation(project(":learning:learning-module-rules"))
     implementation(project(":bpmn-process"))
     implementation(project(":transaction-importer:transaction-importer-api"))
     implementation(project(":transaction-importer:transaction-importer-csv"))
