+++ fintrack-api/src/main/java/com/jongsoft/finance/rest/transaction/TransactionSuggestionResource.java
@@ -1,14 +1,9 @@
 package com.jongsoft.finance.rest.transaction;
 
 import com.jongsoft.finance.core.RuleColumn;
-import com.jongsoft.finance.core.exception.StatusException;
-import com.jongsoft.finance.domain.account.Account;
-import com.jongsoft.finance.rest.model.AccountResponse;
-import com.jongsoft.finance.rest.model.TagResponse;
-import com.jongsoft.finance.rule.RuleDataSet;
-import com.jongsoft.finance.rule.RuleEngine;
+import com.jongsoft.finance.learning.SuggestionEngine;
+import com.jongsoft.finance.learning.SuggestionInput;
 import com.jongsoft.finance.security.AuthenticationRoles;
-import com.jongsoft.lang.Control;
 import io.micronaut.http.annotation.Body;
 import io.micronaut.http.annotation.Controller;
 import io.micronaut.http.annotation.Post;
@@ -16,18 +11,20 @@ import io.micronaut.security.annotation.Secured;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
+import java.time.LocalDate;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 
 @Tag(name = "Transactions")
 @Controller("/api/transactions/suggestions")
 @Secured(AuthenticationRoles.IS_AUTHENTICATED)
 class TransactionSuggestionResource {
 
-    private final RuleEngine ruleEngine;
+    private final SuggestionEngine suggestionEngine;
 
-    TransactionSuggestionResource(RuleEngine ruleEngine) {
-        this.ruleEngine = ruleEngine;
+    TransactionSuggestionResource(SuggestionEngine suggestionEngine) {
+        this.suggestionEngine = suggestionEngine;
     }
 
     @Post
@@ -36,25 +33,27 @@ class TransactionSuggestionResource {
             summary = "Suggest changes",
             description = "Suggest changes to a transaction based upon the rules in the system.")
     Map<String, ?> suggest(@Body TransactionForSuggestionRequest request) {
-        var ruleDataset = new RuleDataSet();
-        Control.Option(request.amount()).ifPresent(value -> ruleDataset.put(RuleColumn.AMOUNT, value));
-        Control.Option(request.source()).ifPresent(value -> ruleDataset.put(RuleColumn.SOURCE_ACCOUNT, value));
-        Control.Option(request.destination()).ifPresent(value -> ruleDataset.put(RuleColumn.TO_ACCOUNT, value));
-        Control.Option(request.description()).ifPresent(value -> ruleDataset.put(RuleColumn.DESCRIPTION, value));
 
-        var suggestions = new HashMap<String, Object>();
-        for (var suggestion : ruleEngine.run(ruleDataset).entrySet()) {
-            var outputSuggestion = switch (suggestion.getValue()) {
-                case Account account -> new AccountResponse(account);
-                // This is the case for category, budget
-                case String value -> value;
-                case com.jongsoft.finance.domain.transaction.Tag tag -> new TagResponse(tag);
+        var transactionInput = new SuggestionInput(
+                LocalDate.now(),
+                request.description(),
+                request.source(),
+                request.destination(),
+                Optional.ofNullable(request.amount()).orElse(0D));
+        var suggestions = suggestionEngine.makeSuggestions(transactionInput);
 
-                default -> throw StatusException.internalError("Could not convert suggestion correctly for type " + suggestion.getValue().getClass());
-            };
+        var output = new HashMap<String, Object>();
 
-            suggestions.put(suggestion.getKey().toString(), outputSuggestion);
+        if (suggestions.budget() != null) {
+            output.put(RuleColumn.BUDGET.toString(), suggestions.budget());
         }
-        return suggestions;
+        if (suggestions.tags() != null) {
+            output.put(RuleColumn.TAGS.toString(), suggestions.tags().stream().map(TagCreateRequest::new).toList());
+        }
+        if (suggestions.category() != null) {
+            output.put(RuleColumn.CATEGORY.toString(), suggestions.category());
+        }
+
+        return output;
     }
 }
