+++ fintrack-api/src/test/java/com/jongsoft/finance/rest/transaction/TransactionSuggestionResourceTest.java
@@ -2,9 +2,10 @@ package com.jongsoft.finance.rest.transaction;
 
 import com.jongsoft.finance.core.RuleColumn;
 import com.jongsoft.finance.core.RuleOperation;
-import com.jongsoft.finance.domain.account.Account;
 import com.jongsoft.finance.domain.transaction.TransactionRule;
+import com.jongsoft.finance.domain.user.Category;
 import com.jongsoft.finance.providers.AccountProvider;
+import com.jongsoft.finance.providers.CategoryProvider;
 import com.jongsoft.finance.providers.TransactionRuleProvider;
 import com.jongsoft.finance.rest.TestSetup;
 import com.jongsoft.lang.Collections;
@@ -16,16 +17,25 @@ import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
+import static org.mockito.Mockito.when;
+
 class TransactionSuggestionResourceTest extends TestSetup {
 
     @Inject
     private TransactionRuleProvider ruleProvider;
     @Inject
-    private AccountProvider accountProvider;
+    private CategoryProvider categoryProvider;
 
     @MockBean
     TransactionRuleProvider ruleProvider() {
         return Mockito.mock(TransactionRuleProvider.class);
+    }
+
+    @MockBean
+    CategoryProvider categoryProvider() {
+        var mock = Mockito.mock(CategoryProvider.class);
+        when(mock.supports(Category.class)).thenReturn(true);
+        return mock;
     }
 
     @MockBean
@@ -47,10 +57,10 @@ class TransactionSuggestionResourceTest extends TestSetup {
                 .build();
         transactionRule.new Condition(1L, RuleColumn.TO_ACCOUNT, RuleOperation.CONTAINS, "Store");
         transactionRule.new Condition(2L, RuleColumn.AMOUNT, RuleOperation.LESS_THAN, "100.00");
-        transactionRule.new Change(1L, RuleColumn.TO_ACCOUNT, "2");
+        transactionRule.new Change(1L, RuleColumn.CATEGORY, "2");
 
-        Mockito.when(accountProvider.lookup(2)).thenReturn(Control.Option(Account.builder().name("Walmart Total Store").id(1L).build()));
-        Mockito.when(ruleProvider.lookup()).thenReturn(Collections.List(transactionRule));
+        when(categoryProvider.lookup(2)).thenReturn(Control.Option(Category.builder().label("Shopping").id(1L).build()));
+        when(ruleProvider.lookup()).thenReturn(Collections.List(transactionRule));
 
         requestSpecification.given()
                 .body("""
@@ -62,6 +72,6 @@ class TransactionSuggestionResourceTest extends TestSetup {
                 .post("/api/transactions/suggestions")
             .then()
                 .statusCode(200)
-                .body("TO_ACCOUNT.name", Matchers.equalTo("Walmart Total Store"));
+                .body("CATEGORY", Matchers.equalTo("Shopping"));
     }
 }
