--- learning/language-engine/src/main/java/com/jongsoft/finance/llm/agent/AiClassification.java
@@ -1,6 +1,0 @@
-package com.jongsoft.finance.llm.agent;
-
-import java.util.List;
-
-public record AiClassification(String budget, String category, List<String> tags) {
-}
