+++ learning/learning-module-llm/src/main/java/com/jongsoft/finance/llm/AiEnabled.java
@@ -1,0 +1,7 @@
+package com.jongsoft.finance.llm;
+
+import io.micronaut.context.annotation.Requires;
+
+@Requires(env = "ai")
+public @interface AiEnabled {
+}
