+++ learning/learning-module-llm/src/main/java/com/jongsoft/finance/llm/tools/AiTool.java
