+++ learning/learning-module-llm/src/main/resources/application-ai.yaml
