+++ learning/learning-module-llm/src/test/resources/en/sample-transactions.csv
