+++ learning/learning-module-llm/src/test/resources/en/test-config.properties
