+++ learning/learning-module-llm/src/test/resources/nl/test-config.properties
