+++ learning/learning-module-rules/src/main/java/com/jongsoft/finance/automation/ScheduleTerminateContractHandler.java
