+++ learning/learning-module-rules/src/main/java/com/jongsoft/finance/rule/locator/NoopLocator.java
