+++ learning/learning-module-rules/src/main/java/com/jongsoft/finance/rule/matcher/StringMatcher.java
