+++ learning/learning-module-rules/src/test/java/com/jongsoft/finance/automation/ScheduleTerminateContractCommandHandlerTest.java
