+++ learning/learning-module/build.gradle.kts
@@ -1,0 +1,15 @@
+micronaut {
+    testRuntime("junit5")
+}
+
+java {
+    modularity.inferModulePath.set(true)
+}
+
+dependencies {
+    implementation(project(":core"))
+    implementation(project(":domain"))
+
+    testRuntimeOnly(mn.logback.classic)
+    testImplementation(libs.bundles.junit)
+}
