+++ learning/learning-module/src/main/java/com/jongsoft/finance/learning/SuggestionEngine.java
@@ -1,0 +1,7 @@
+package com.jongsoft.finance.learning;
+
+public interface SuggestionEngine {
+
+    SuggestionResult makeSuggestions(SuggestionInput transactionInput);
+
+}
