+++ learning/learning-module/src/main/java/com/jongsoft/finance/learning/SuggestionInput.java
@@ -1,0 +1,6 @@
+package com.jongsoft.finance.learning;
+
+import java.time.LocalDate;
+
+public record SuggestionInput(LocalDate date, String description, String fromAccount, String toAccount, double amount) {
+}
