+++ learning/learning-module/src/main/java/com/jongsoft/finance/learning/SuggestionResult.java
@@ -1,0 +1,6 @@
+package com.jongsoft.finance.learning;
+
+import java.util.List;
+
+public record SuggestionResult(String budget, String category, List<String> tags) {
+}
